+++ lib/src/grammar.dart
@@ -2,13 +2,13 @@ part of angel_route.src.router;
 
 class RouteGrammar {
   static final Parser<String> notSlash =
-      match(new RegExp(r'[^/]+')).value((r) => r.span.text);
+      match<String>(new RegExp(r'[^/]+')).value((r) => r.span.text);
 
   static final Parser<RegExp> regExp = match<RegExp>(new RegExp(r'\((.+)\)'))
       .value((r) => new RegExp(r.scanner.lastMatch[1]));
 
   static final Parser<String> parameterName =
-      match(new RegExp(r':([A-Za-z0-9_]+)'))
+      match<String>(new RegExp(r':([A-Za-z0-9_]+)'))
           .value((r) => r.span.text.substring(1));
 
   static final Parser<ParameterSegment> parameterSegment = chain([
@@ -16,7 +16,7 @@ class RouteGrammar {
     match('?').value((r) => true).opt(),
     regExp.opt(),
   ]).map((r) {
-    var s = new ParameterSegment(r.value[0], r.value[2]);
+    var s = new ParameterSegment(r.value[0] as String, r.value[2] as RegExp);
     return r.value[1] == true ? new OptionalSegment(s) : s;
   });
 
@@ -26,16 +26,17 @@ class RouteGrammar {
         .map((r) => r.span.text),
     parameterSegment,
   ]).map((r) {
-    return new ParsedParameterSegment(r.value[0], r.value[1]);
+    return new ParsedParameterSegment(
+        r.value[0] as String, r.value[1] as ParameterSegment);
   });
 
   static final Parser<WildcardSegment> wildcardSegment =
-      match('*').value((r) => new WildcardSegment());
+      match<WildcardSegment>('*').value((r) => new WildcardSegment());
 
   static final Parser<ConstantSegment> constantSegment =
-      notSlash.map((r) => new ConstantSegment(r.value));
+      notSlash.map<ConstantSegment>((r) => new ConstantSegment(r.value));
 
-  static final Parser<RouteSegment> routeSegment = any([
+  static final Parser<RouteSegment> routeSegment = any<RouteSegment>([
     parsedParameterSegment,
     parameterSegment,
     wildcardSegment,
@@ -62,7 +63,8 @@ class RouteDefinition {
       if (out == null)
         out = s.compile(isLast);
       else
-        out = s.compileNext(out.then(match('/')).index(0), isLast);
+        out = s.compileNext(
+            out.then(match('/')).index(0).cast<Map<String, dynamic>>(), isLast);
     }
 
     return out;
@@ -88,13 +90,13 @@ class ConstantSegment extends RouteSegment {
 
   @override
   Parser<Map<String, dynamic>> compile(bool isLast) {
-    return match(text).value((r) => {});
+    return match<Map<String, dynamic>>(text).value((r) => <String, dynamic>{});
   }
 
   @override
   Parser<Map<String, dynamic>> compileNext(
       Parser<Map<String, dynamic>> p, bool isLast) {
-    return p.then(compile(isLast)).index(0);
+    return p.then(compile(isLast)).index(0).cast<Map<String, dynamic>>();
   }
 }
 
@@ -117,7 +119,7 @@ class WildcardSegment extends RouteSegment {
   @override
   Parser<Map<String, dynamic>> compileNext(
       Parser<Map<String, dynamic>> p, bool isLast) {
-    return p.then(_compile(isLast)).index(0);
+    return p.then(_compile(isLast)).index(0).cast<Map<String, dynamic>>();
   }
 }
 
@@ -140,8 +142,9 @@ class OptionalSegment extends ParameterSegment {
   Parser<Map<String, dynamic>> compileNext(
       Parser<Map<String, dynamic>> p, bool isLast) {
     return p.then(_compile().opt()).map((r) {
-      if (r.value[1] == null) return r.value[0];
-      return r.value[0]..addAll({name: Uri.decodeComponent(r.value[1])});
+      if (r.value[1] == null) return r.value[0] as Map<String, dynamic>;
+      return (r.value[0] as Map<String, dynamic>)
+        ..addAll({name: Uri.decodeComponent(r.value[1] as String)});
     });
   }
 }
@@ -158,22 +161,24 @@ class ParameterSegment extends RouteSegment {
     return 'Param: $name';
   }
 
-  Parser<Map<String, dynamic>> _compile() {
+  Parser<String> _compile() {
     return regExp != null
-        ? match(regExp).value((r) => r.span.text)
+        ? match<String>(regExp).value((r) => r.span.text)
         : RouteGrammar.notSlash;
   }
 
   @override
   Parser<Map<String, dynamic>> compile(bool isLast) {
-    return _compile().map((r) => {name: Uri.decodeComponent(r.span.text)});
+    return _compile().map<Map<String, dynamic>>(
+        (r) => {name: Uri.decodeComponent(r.span.text)});
   }
 
   @override
   Parser<Map<String, dynamic>> compileNext(
       Parser<Map<String, dynamic>> p, bool isLast) {
     return p.then(_compile()).map((r) {
-      return r.value[0]..addAll({name: Uri.decodeComponent(r.value[1])});
+      return (r.value[0] as Map<String, dynamic>)
+        ..addAll({name: Uri.decodeComponent(r.value[1] as String)});
     });
   }
 }
@@ -205,8 +210,10 @@ class ParsedParameterSegment extends RouteSegment {
   Parser<Map<String, dynamic>> compileNext(
       Parser<Map<String, dynamic>> p, bool isLast) {
     return p.then(parameter._compile()).map((r) {
-      return r.value[0]
-        ..addAll({parameter.name: getValue(Uri.decodeComponent(r.value[1]))});
+      return (r.value[0] as Map<String, dynamic>)
+        ..addAll({
+          parameter.name: getValue(Uri.decodeComponent(r.value[1] as String))
+        });
     });
   }
 }