+++ lib/src/symlink_route.dart
@@ -2,7 +2,7 @@ part of angel_route.src.router;
 
 /// Placeholder [Route] to serve as a symbolic link
 /// to a mounted [Router].
-class SymlinkRoute extends Route {
-  final Router router;
+class SymlinkRoute<T> extends Route<T> {
+  final Router<T> router;
   SymlinkRoute(String path, this.router) : super(path, method: null, handlers: null);
 }